Reorganized file io functions to live in utils
